Move `yarn` logic into `NpmDiscovery`
