+++ app/models/shipit/deploy_spec/npm_discovery.rb
@@ -8,7 +8,11 @@ module Shipit
       end
 
       def discover_package_json
-        npm_install if npm?
+        (yarn_install if yarn?) || (npm_install if npm?)
+      end
+
+      def yarn_install
+        [%(yarn install --no-progress)]
       end
 
       def npm_install
@@ -16,7 +20,12 @@ module Shipit
       end
 
       def discover_review_checklist
-        discover_npm_checklist || super
+        discover_yarn_checklist || discover_npm_checklist || super
+      end
+
+      def discover_yarn_checklist
+        [%(<strong>Don't forget version and tag before publishing!</strong> You can do this with:<br/>
+          yarn version --new-version <strong>&lt;major|minor|patch&gt;</strong> && git push --tags</pre>)] if yarn?
       end
 
       def discover_npm_checklist
@@ -39,12 +48,23 @@ module Shipit
         file('package.json')
       end
 
-      def discover_npm_package
-        publish_npm_package if npm?
+      def yarn?
+        yarn_lock.exist? && public?
+      end
+
+      def yarn_lock
+        file('./yarn.lock')
       end
 
       def discover_deploy_steps
-        discover_npm_package || super
+        (publish_yarn_package if yarn?) || (publish_npm_package if npm?) || super
+      end
+
+      def publish_yarn_package
+        check_tags = 'assert-npm-version-tag'
+        publish = 'yarn publish'
+
+        [check_tags, publish]
       end
 
       def publish_npm_package
